requirements.txt

Django==4.1.7
Pillow==9.4.0

